pyproject.toml
[tool.poetry]
name = "secret-scanner"
version = "0.1.0"
description = "A production-grade secret scanner for Git repositories"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "secret_scanner", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
gitpython = "^3.1.40"
pyyaml = "^6.0.1"
colorama = "^0.4.6"
streamlit = "^1.31.0"
pandas = "^2.2.0"
plotly = "^5.18.0"
pygithub = "^2.1.1"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.12.1"
ruff = "^0.1.9"

[tool.poetry.scripts]
secret-scanner = "secret_scanner.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
